requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.10.0
Flask-Limiter==3.5.0
Flask-Migrate==4.1.0
email-validator==2.1.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
redis==5.0.1
PyJWT==2.8.0
